export the sold column #375
