+++ frontend/src/components/products/ProductCategoryConfig.vue
@@ -28,8 +28,6 @@ import { deleteProductCategory } from "../../requests/productCategory";
 import { useUiFeedback } from "../../store/uiFeedbackStore";
 import { useConfigStore } from "../../store/configStore";
 import { useProductStore } from "../../store/productStore";
-import { sanitizeFileName } from "@lebenswurzel/solawi-bedarf-shared/src/util/fileHelper.ts";
-import { objectToCsv } from "@lebenswurzel/solawi-bedarf-shared/src/pdf/overviewPdfs.ts";
 import { ProductCategoryType } from "@lebenswurzel/solawi-bedarf-shared/src/enum.ts";
 const t = language.pages.product;
 
@@ -50,24 +48,6 @@ const dialogProductCategory = ref<ProductCategory>({
 const onEditProductCategory = () => {
   dialogProductCategory.value = { ...props.productCategoryWithProducts };
   openProductCategory.value = true;
-};
-
-const onExportProducts = async () => {
-  try {
-    const csv = objectToCsv(props.productCategoryWithProducts.products);
-    const blob = new Blob([csv], { type: "text/csv" });
-    const url = window.URL.createObjectURL(blob);
-    const a = document.createElement("a");
-    a.href = url;
-    a.download = sanitizeFileName(props.productCategoryWithProducts.name);
-    document.body.appendChild(a);
-    a.click();
-    document.body.removeChild(a);
-    window.URL.revokeObjectURL(url);
-  } catch (e) {
-    uiFeedback.setError("" + e);
-    throw e;
-  }
 };
 
 const onDelete = () => {
@@ -124,13 +104,6 @@ const onCloseProductCategory = async () => {
             <v-list-item-title>
               <v-btn @click="onEditProductCategory">
                 {{ language.app.actions.edit }}
-              </v-btn>
-            </v-list-item-title>
-          </v-list-item>
-          <v-list-item>
-            <v-list-item-title>
-              <v-btn @click="onExportProducts" color="secondary">
-                CSV-Export
               </v-btn>
             </v-list-item-title>
           </v-list-item>
